requirements.txt
authlib==1.0
django==4.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.4
python-dotenv==0.19
requests==2.27
whitenoise==6.2.0